# api/openapi.yaml
openapi: 3.1.0
...
info:
  title: Alexandryn system API
  version: 0.0.0
  description: >-
    The contract between the Go server and every client (the Electron
    renderer and any LAN browser). Errors share one shape across every
    endpoint.
servers:
  - url: /
    description: The Go server's own origin (loopback by default, constitution §6).

# Path-based versioning. Health lives at the root, outside the versioned
# surface, because it predates and is independent of it.
paths:
...
  /api/v1/works/{id}:
    get:
      operationId: getWork
      summary: Get a single work's detail
      description: >-
        Returns one Work's metadata, owned editions (with formats), and
        collection memberships. Absent-but-optional fields (cover URL,
        availability) are simply missing from the response object, not
        present-but-null.
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: WorkID (internal). Malformed shape → 400; not found → 404.
      responses:
        '200':
          description: The work detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkDetail'
              example:
                id: "01JXXXXXXXXXXXXXXXXXXXXXXX"
                title: "Middlemarch"
                subtitle: ""
                authors: ["George Eliot"]
                subjects: ["Fiction"]
                originalLanguage: "en"
                ownedEditions:
                  - id: "01JXXXXXXXXXXXXXXXXXXXXXXY"
                    language: "en"
                    isbn: "9780141439549"
                    publisher: "Penguin Classics"
                    publicationYear: 2003
                    addedAt: "2026-01-15T10:00:00Z"
                    formats: ["epub"]
                collections:
                  - id: "01JXXXXXXXXXXXXXXXXXXXXXXZ"
                    name: "Classics"
                    addedAt: "2026-01-20T08:00:00Z"
        '400':
          description: Malformed WorkID (not a valid ID shape).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: invalid_input
                message: "id: not a valid work ID"
                correlationId: 00000000-0000-0000-0000-000000000000
        '404':
          description: No work with that ID exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: not_found
                message: "no work with that id"
                correlationId: 00000000-0000-0000-0000-000000000000
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required: [code, message, correlationId]
      description: >-
        The one error shape every endpoint uses. `correlationId` matches
        the id in the server's structured logs for the same request.
      properties:
        code:
          type: string
          description: Machine-readable category (domain.Category).
          example: not_found
        message:
          type: string
          description: >-
            Human-readable, specific, no apology, says what to do next
            (constitution §11).
          example: no such endpoint
        correlationId:
          type: string
          description: Ties this response to one server log line.
          example: 00000000-0000-0000-0000-000000000000

    # ── Library domain schemas ───────────────────────────────────────
...
    OwnedEdition:
      type: object
      additionalProperties: false
      required: [id, language, publisher, formats]
      description: >-
        One owned Edition in a WorkDetail response. formats
        carries the distinct Format values across every SourceOffering
        referencing this Edition.
      properties:
        id:
          type: string
          description: EditionID (internal).
        language:
          type: string
          description: BCP-47 language tag.
        isbn:
          type: string
          nullable: true
        publisher:
          type: string
          description: Empty string when absent.
        publicationYear:
          type: integer
          nullable: true
        addedAt:
          type: string
          format: date-time
          description: LibraryEntry.added_at for this Edition.
        formats:
          type: array
          items:
            type: string
          description: >-
            Distinct Format values across every SourceOffering referencing
            this Edition. Empty array when no SourceOfferings exist. Always
            present (never absent).
    CollectionRef:
      type: object
      additionalProperties: false
      required: [id, name]
      description: >-
        Minimal collection reference embedded in WorkSummary and
        WorkDetail.
      properties:
        id:
          type: string
          description: CollectionID (internal).
        name:
          type: string
        addedAt:
          type: string
          format: date-time
          description: >-
            Per-membership added-at timestamp — when this Work was added
            to this Collection.
    WorkDetail:
      type: object
      additionalProperties: false
      required: [id, title, subtitle, authors, ownedEditions, collections]
      description: >-
        Response body for GET /api/v1/works/{id}.
      properties:
        id:
          type: string
        title:
          type: string
        subtitle:
          type: string
          description: Empty string when absent.
        authors:
          type: array
          items:
            type: string
          description: Author display names.
        subjects:
          type: array
          items:
            type: string
        originalLanguage:
          type: string
          nullable: true
          description: BCP-47 language tag.
        ownedEditions:
          type: array
          items:
            $ref: '#/components/schemas/OwnedEdition'
          description: >-
            Editions of this Work for which a LibraryEntry exists.
        collections:
          type: array
          items:
            $ref: '#/components/schemas/CollectionRef'
          description: >-
            Collections this Work is a member of, each with the
            per-membership added-at timestamp.
...
security:
  - bearerAuth: []